refactor(phase4c): checkpoint 保存系を library/checkpoint_io.py に分離
train_util.py から以下を library/checkpoint_io.py に移動:
- ファイル名テンプレート定数 (EPOCH_FILE_NAME / STEP_STATE_NAME 等 10 個)
- default_if_none / get_epoch_ckpt_name / get_step_ckpt_name /
  get_last_ckpt_name (チェックポイント名ビルダ)
- get_remove_epoch_no / get_remove_step_no
  (--save_last_n_epochs / --save_last_n_steps の rotate 計算)
- save_sd_model_on_epoch_end_or_stepwise(_common) /
  save_sd_model_on_train_end(_common)
  (SD 1.x/2.x checkpoint 保存 + HF Hub upload + rotate)
- save_and_remove_state_on_epoch_end / save_and_remove_state_stepwise /
  save_state_on_train_end (accelerator state 保存 + rotate)

train_util.py は re-export shim を維持。当リポジトリ内 14 トップレベル
スクリプトと library/{anima,flux,leco,lumina,sd3,sdxl}_train_util.py 6 ファイル
の参照を checkpoint_io.* (alias) に切替。

train_util.py: 1,275 → 975 行 (250 行削減)。
library/checkpoint_io.py: 348 行 (新規)。
