+++ lumina_train.py
@@ -38,6 +38,7 @@ from library.sd3_train_utils import FlowMatchEulerDiscreteScheduler
 import library.train_util as train_util
 import library.logging_util as logging_util
 import library.loss as loss_util
+import library.checkpoint_io as checkpoint_io
 
 from library.utils import setup_logging, add_logging_arguments
 
@@ -897,7 +898,7 @@ def train(args):
     optimizer_eval_fn()
 
     if args.save_state or args.save_state_on_train_end:
-        train_util.save_state_on_train_end(args, accelerator)
+        checkpoint_io.save_state_on_train_end(args, accelerator)
 
     del accelerator  # この後メモリを使うのでこれは消す
