+++ sdxl_train.py
@@ -22,6 +22,7 @@ from library import deepspeed_utils, sdxl_model_util, strategy_base, strategy_sd
 import library.train_util as train_util
 import library.logging_util as logging_util
 import library.loss as loss_util
+import library.checkpoint_io as checkpoint_io
 
 from library.utils import setup_logging, add_logging_arguments
 
@@ -879,7 +880,7 @@ def train(args):
     accelerator.end_training()
 
     if args.save_state or args.save_state_on_train_end:
-        train_util.save_state_on_train_end(args, accelerator)
+        checkpoint_io.save_state_on_train_end(args, accelerator)
 
     del accelerator  # この後メモリを使うのでこれは消す
